Build mod files into .\export\mod\
Also moved mod functions into a helper file.
