+++ basic_armors.py
@@ -3,35 +3,9 @@ import sys
 import shutil
 import copy
 import re
-from helpers.create_db import create_db
 from bs4 import BeautifulSoup, Tag, NavigableString
-
-def write_db(filepath, fg_data):
-    with open(filepath, mode='w', encoding='UTF-8', errors='strict', buffering=1) as file:
-        # item counter
-        id = 0
-
-        file.write('<root version="2.9">\n')
-        file.write('\t<item>\n')
-        for entry in fg_data:
-            id += 1
-            entrry_id = "00000"[0:len("00000")-len(str(id))] + str(id)
-            file.write(f'\t\t<id-{entrry_id}>\n')
-
-            for tag in sorted(entry.keys()):
-                if entry[tag] != '':
-                    # choose the write statement depending on the 'type' attribute
-                    if re.search('^(powers|props)$', tag):
-                        file.write(f'\t\t\t<{tag}>{entry[tag]}</{tag}>\n')
-                    elif re.search('^(ac|bonus|checkpenalty|level|min_enhance|profbonus|range|speed|weight)$', tag):
-                        file.write(f'\t\t\t<{tag} type="number">{entry[tag]}</{tag}>\n')
-                    else:
-                        file.write(f'\t\t\t<{tag} type="string">{entry[tag]}</{tag}>\n')
-
-            file.write(f'\t\t</id-{entrry_id}>\n')
-        file.write('\t</item>\n')
-        file.write('</root>')
-        return str(id)
+from helpers.create_db import create_db
+from helpers.mod_helpers import write_db
 
 if __name__ == '__main__':
 
@@ -55,7 +29,7 @@ if __name__ == '__main__':
         input('Press enter to close.')
         sys.exit(0)
 
-    for i, row in enumerate(db):
+    for i, row in enumerate(db, start=1):
 
         # Parse the HTML text 
         html = row['Txt']
@@ -173,14 +147,14 @@ if __name__ == '__main__':
     print(str(write_id) + " module entries written. Job done.")
 
     try:
-        os.remove('export/basic_armors/4E_Basic_Armors.mod')
+        os.remove('export/mods/4E_Basic_Armors.mod')
     except FileNotFoundError:
         print("Cleanup not needed.")
     try:
-        shutil.make_archive('export/basic_armors/4E_Basic_Armors', 'zip', 'export/basic_armors/data/')
-        os.rename('export/basic_armors/4E_Basic_Armors.zip', 'export/basic_armors/4E_Basic_Armors.mod')
+        shutil.make_archive('export/mods/4E_Basic_Armors', 'zip', 'export/basic_armors/data/')
+        os.rename('export/mods/4E_Basic_Armors.zip', 'export/mods/4E_Basic_Armors.mod')
         print("\nDatabase added and module generated!")
-        print("You can find it in the 'export\\basic_armors' folder\n")
+        print("You can find it in the 'export\\mods' folder\n")
     except Exception as e:
         print(f"Error creating zipped .mod file:\n{e}")
         print("\nManually zip the contents of the 'export\\basic_armor\\data' folder to create the mod.")
